Remove dead code, organize packages
