+++ app/src/main/AndroidManifest.xml
@@ -37,7 +37,7 @@
             android:name="nl.tudelft.trustchain.debug.DebugActivity"
             android:parentActivityName=".ui.dashboard.DashboardActivity" />
 
-        <activity android:name="com.example.musicdao.net.MusicService" />
+        <activity android:name="com.example.musicdao.MusicService" />
 
         <service android:name="nl.tudelft.trustchain.app.service.TrustChainService" />
     </application>
